Add optional grounding frames
This option forces the robot frame in-between the specified grounding
frames to lie on the ground plane. This is done by shifting all segment
poses by the computed offset between this grounding frame and the floor
plane.

The need for it arises from the problem of synchronizing a tracked human
with its simulated counterpart in MuJoCo.

As-is this works only for the main robot.